Refactoring of ugrid_dataset. Start wrapping before runtime.
